requirements.txt
# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
playwright>=1.40.0  # For JavaScript-heavy sites

# RAG System - Vector DB & Embeddings
chromadb==0.4.22
sentence-transformers==2.5.1

# LLM APIs (choose one or both)
openai==1.12.0          # For GPT-4o-mini (fast & cheap)
anthropic>=0.70.0       # For Claude (accurate & smart)
tiktoken==0.6.0         # Token counting for OpenAI

# Telegram Bot & Automation
python-telegram-bot>=22.0  # For Telegram integration
python-dotenv>=1.0.0       # For .env configuration management
schedule>=1.2.0            # For automated task scheduling

